Created DollarViewModel / Init ViewModel in ContentView / Changed model in currency from Decimal to Double / Added extension to Double to fit currency display

+++ dolaruco-app/Models/Currency.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 struct Currency: Decodable {
-    let compra: Decimal
-    let venta: Decimal
+    let compra: Double
+    let venta: Double
     let casa: String
     let nombre: String
     let moneda: String
